remove deleteAccountModal, add Alert in setting.tsx, error fixed, API and delete accout feature working well
